test(runtime): ME-R7 auto-compilacion del runtime en clon limpio + deudas D1/D2/D6
Causa E (Manual 9 S9.7: tests reales en instalacion limpia) + deudas aprobadas por el Arquitecto.
- tests/conftest.py: fixture autouse session (HARNESS, sin tocar tests) que auto-compila desde fuente los .o del runtime a la raiz (tweetnacl/synapse_rt/synapse_rt_memory/-DSYNAPSE_DEBUG_MEM/synapse_rt_concurrency/axon_rt) con dependencias mtime (fuente + headers incl. embedded_libs.h) y linkea 5 binarios de test (work_stealing, cluster_raft, path_traversal_new, ed25519_axon_new con todos los .o; gen_axon_test_fixtures solo con tweetnacl.o por su randombytes propio). Post-review: .o parcial se borra en fallo; fallo no aborta suite (WARNING).
- D1: tr_grabar_snapshot long long -> int (alinea contrato std/debug.syn entero) + prototipos colgantes en test_distributed_debug.c y test_memory_snapshots.c.
- D2: test_axon_e2e_native.c y extension aprobada test_path_traversal.c -> salida .axon_cache/tar_test_out (antes borraban fixture trackeado tests/fixtures/tar_test_out/); entrada malicious.tar intacta; cero cambios de asercion.
- D6: cli.py _compilar_runtime_sanitizado(dir_rel, san_flags) compila runtime modular desde fuente con -fsanitize a build/obj/; auditorias memoria/hilos lo usan (antes .o pre-ME-R2).
Validado: clon limpio sin .o ni binarios -> pytest test_axon_e2e.py 5 passed, integracion 331 passed; suite completa 667 passed, 0 failed; fixture trackeado intacto tras corridas.

+++ synapse_rt.c
@@ -5785,7 +5785,7 @@ int tr_grabar_bifurcacion(int linea, int rama, CadenaSegura id_funcion) {
 // valor_entero: integer value (or 0 if using texto)
 // valor_texto: string value (or empty if using entero)
 // linea: source line number
-int tr_grabar_snapshot(CadenaSegura nombre_variable, long long valor_entero,
+int tr_grabar_snapshot(CadenaSegura nombre_variable, int valor_entero,
                        CadenaSegura valor_texto, int linea) {
     if (!_tr_initialized) return -1;
     int seq = _tr_next_seq();
