+++ tests/test_distributed_debug.c
@@ -34,7 +34,7 @@ extern int tr_grabar_llamada(CadenaSegura funcion, int linea, int num_args);
 extern int tr_grabar_retorno(CadenaSegura funcion, int linea);
 extern int tr_grabar_bifurcacion(int linea, int rama, CadenaSegura id_funcion);
 extern int tr_grabar_error(CadenaSegura mensaje, int linea);
-extern int tr_grabar_snapshot(CadenaSegura nombre_variable, long long valor_entero,
+extern int tr_grabar_snapshot(CadenaSegura nombre_variable, int valor_entero,
                               CadenaSegura valor_texto, int linea);
 extern int tr_total_eventos(void);
